Gives each of the panels in the cosmic ray tab its own file
